requirements.txt
fastapi
requests
python-dotenv
jinja2
airtable